都内の滞在人口の増減状況: Rename components/StayingPopulation.vue to components/index/SiteTopUpper/StayingPopulation.vue
